+++ lib/index.ts
@@ -1,31 +1,32 @@
 import type { GPTProviderProps } from "./components/GPTProvider/GPTProvider.type";
 import type { GPTSlotProps } from "./components/GPTSlot/GPTSlot.type";
 import type {
-	Size,
-	Sizes,
-	SlotLoadEvent,
-	SlotRequestEvent,
-	SlotRenderEndedEvent,
-	SlotViewableEvent,
-	RewardedOnReadyEvent,
-	RewardedOnGrantedEvent,
+  Size,
+  Sizes,
+  SlotLoadEvent,
+  SlotRequestEvent,
+  SlotRenderEndedEvent,
+  SlotViewableEvent,
+  RewardedOnReadyEvent,
+  RewardedOnGrantedEvent,
 } from "./types";
 
 import { GPTProvider } from "./components/GPTProvider/GPTProvider";
 import { GPTSlot } from "./components/GPTSlot/GPTSlot";
 import { GUIDELINES } from "./constants/guidelines";
 import { useGPT } from "./hooks/useGPT/useGPT";
+export * from "./utils/utils";
 
 export type {
-	GPTSlotProps,
-	GPTProviderProps,
-	SlotLoadEvent,
-	SlotRequestEvent,
-	SlotViewableEvent,
-	SlotRenderEndedEvent,
-	RewardedOnReadyEvent,
-	RewardedOnGrantedEvent,
-	Sizes,
-	Size,
+  GPTSlotProps,
+  GPTProviderProps,
+  SlotLoadEvent,
+  SlotRequestEvent,
+  SlotViewableEvent,
+  SlotRenderEndedEvent,
+  RewardedOnReadyEvent,
+  RewardedOnGrantedEvent,
+  Sizes,
+  Size,
 };
 export { GPTProvider, GPTSlot, useGPT, GUIDELINES };